[refactor] 디렉토리 구조 개편, GatheringsHeader에 subTitle, title 수정
