+++ src/app/auth/signin/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next';
-import AuthPoster from '@/components/auth/shared/AuthPoster';
+import AuthPoster from '@/components/auth/AuthPoster';
 import SignInForm from '@/components/auth/SignInForm';
 
 export const metadata: Metadata = {
