+++ src/components/gatherings/GatheringsList.tsx
@@ -9,7 +9,7 @@ import { UserRoundCheck } from "lucide-react";
 import { filterGatheringsByType, getUniqueGatherings } from '@/utils/gatherings/gatheringsUtils';
 import ImageWithFallback from "@/components/shared/ImageWithFallback";
 import JoinedCountsProgressBar from "@/components/gatherings/shared/JoinedCountsProgressBar";
-import SaveToggleButton from '@/components/gatherings/shared/SaveToggleButton';
+import SaveToggleButton from '@/components/shared/SaveToggleButton';
 import DateReminder from '@/components/shared/DateReminder';
 
 // 스타일 상수
