+++ src/utils/gatherings/createGatheringSchema.ts
